[#109] Refactor: 명확한 패키지 구분을 위한 변경
- common 패키지 : 공통적으로 필요한 패키지를 모아둠.
- common/exception 패키지 : 커스텀 예외 처리 클래스를 모아둠.
- common/response 패키지 : 응답 객체를 모아둠.
- domain 패키지 : entity 클래스와 repository를 모아둠.
