+++ BE/src/main/java/com/codesquad/team1/signup/common/exception/ForbiddenException.java
@@ -1,4 +1,4 @@
-package com.codesquad.team1.signup.Exception;
+package com.codesquad.team1.signup.common.exception;
 
 public class ForbiddenException extends RuntimeException {
