+++ BE/src/main/java/com/codesquad/team1/signup/domain/interest/InterestRepository.java
@@ -1,5 +1,6 @@
-package com.codesquad.team1.signup.repository;
+package com.codesquad.team1.signup.domain.interest;
 
+import com.codesquad.team1.signup.domain.interest.Interest;
 import org.springframework.data.repository.CrudRepository;
 
 public interface InterestRepository extends CrudRepository<Interest, Integer> {}
