+++ BE/src/main/java/com/codesquad/team1/signup/domain/user/UserRepository.java
@@ -1,5 +1,6 @@
-package com.codesquad.team1.signup.repository;
+package com.codesquad.team1.signup.domain.user;
 
+import com.codesquad.team1.signup.domain.user.User;
 import org.springframework.data.jdbc.repository.query.Query;
 import org.springframework.data.repository.CrudRepository;
